src/faiss_services/faiss_requirements.txt
faiss-cpu==1.7.3
pandas==1.4.4
pytest==7.2.0
python-json-logger==2.0.4